Move primitives::sinsemilla::i2lebsp_k into constants::sinsemilla.
